Initial commit of OpenGML
OpenGML provifing 2D and 3D versions to create nested geometries based
on the mathematical Geometric Musical Language.

A common codebase which abstracts to Kivy and PyGame + OpenGL for
the graphics engines for 2D and 3D versions.

Examples are provided for both 2D and 3D variants with minimal coding
required to use either or switch between them.