Linkage with Frequencies
Linkage with just phase resulted in systems where small amounts of
linkage lead to phase locking. The phase was being locked irrespective
of frequency and so not in consideration of time.

Introduction of GML_Bond linkage by both frequency and phase allows for
a more dynamic and generative system.

Examples moved to proivide better scope of grouping.

Common prime number functions moved to a single file.